Enhancement: Allow UI extensibility out of experimental mode #829
