+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - [#819](https://github.com/estruyf/vscode-front-matter/issues/819): Added new extensibility support for media scripts
 - [#822](https://github.com/estruyf/vscode-front-matter/issues/822): Added docs to the panel & dashboard views
+- [#829](https://github.com/estruyf/vscode-front-matter/issues/829): UI extensibility is now generally available
 
 ### 🐞 Fixes
